refactor(SF-135): extract FrontendPluginBase from 4 frontend plugins
claude_frontend, codex_frontend, gemini_frontend, ollama_frontend each
ran ~300 lines of nearly-identical bookkeeping: spec resolution +
caching, a side-spawned FastAPI server, lifecycle hooks. Real
differences were just plugin metadata, settings defaults, the CLI
binary name to check at preflight, and the per-provider router
factory.

Extract to frontend_base/plugin_base.py:

- FrontendSettingsBase: shared pydantic Settings with sensible
  defaults. Subclasses override env_prefix + provider-specific
  defaults (upstream_url, listen_port, default_backend_path) and
  inherit the rest.
- FrontendPluginBase: implements __init__ (cache+lock+app slot),
  _get_spec_names, customize_schema, preflight (CLI binary check
  driven by class attr, with _SF_SUBPROCESS skip preserved),
  _collect_spec_names, _get_spec_urls, get_active_expression,
  resolve_context (the cached double-checked-locking spec resolver),
  _get_backend_url, _fetch_sync, setup (the standalone-FastAPI
  pattern), _run_server, _on_shutdown, teardown, _stop. Subclass
  contract: cli_binary (None to skip), cli_install_hint, app_title,
  create_router.

Per-plugin shrinkage:
- claude_frontend/plugin.py: 363 → 42 lines (-88%)
- codex_frontend/plugin.py:  306 → 44 lines (-86%)
- gemini_frontend/plugin.py: 287 → 43 lines (-85%)
- ollama_frontend/plugin.py: 308 → 50 lines (-84%)
- frontend_base/plugin_base.py: +378 lines (new)

Net: -707 lines.

claude_intercept/tests/test_intercept.py: patch targets updated to
point at frontend_base.plugin_base.uvicorn / _wait_for_port (where
the symbols now live).

722 plugin unit tests pass.

https://app.asana.com/1/1185126988600652/project/1213628819033917/task/1214359276383119
